+++ packages/react/src/components/progress/progress-label.tsx
@@ -4,13 +4,13 @@ import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { useProgressContext } from './use-progress-context'
 
 export interface ProgressLabelBaseProps extends PolymorphicProps {}
-export interface ProgressLabelProps extends HTMLProps<'label'>, ProgressLabelBaseProps {}
+export interface ProgressLabelProps extends HTMLProps<'span'>, ProgressLabelBaseProps {}
 
-export const ProgressLabel = forwardRef<HTMLLabelElement, ProgressLabelProps>((props, ref) => {
+export const ProgressLabel = forwardRef<HTMLSpanElement, ProgressLabelProps>((props, ref) => {
   const progress = useProgressContext()
   const mergedProps = mergeProps(progress.getLabelProps(), props)
 
-  return <ark.label {...mergedProps} ref={ref} />
+  return <ark.span {...mergedProps} ref={ref} />
 })
 
 ProgressLabel.displayName = 'ProgressLabel'