+++ packages/react/src/components/radio-group/radio-group-label.tsx
@@ -4,13 +4,13 @@ import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { useRadioGroupContext } from './use-radio-group-context'
 
 export interface RadioGroupLabelBaseProps extends PolymorphicProps {}
-export interface RadioGroupLabelProps extends HTMLProps<'label'>, RadioGroupLabelBaseProps {}
+export interface RadioGroupLabelProps extends HTMLProps<'span'>, RadioGroupLabelBaseProps {}
 
-export const RadioGroupLabel = forwardRef<HTMLLabelElement, RadioGroupLabelProps>((props, ref) => {
+export const RadioGroupLabel = forwardRef<HTMLSpanElement, RadioGroupLabelProps>((props, ref) => {
   const radioGroup = useRadioGroupContext()
   const mergedProps = mergeProps(radioGroup.getLabelProps(), props)
 
-  return <ark.label {...mergedProps} ref={ref} />
+  return <ark.span {...mergedProps} ref={ref} />
 })
 
 RadioGroupLabel.displayName = 'RadioGroupLabel'