+++ packages/react/src/components/segment-group/segment-group-label.tsx
@@ -5,13 +5,13 @@ import { parts } from './segment-group.anatomy'
 import { useSegmentGroupContext } from './use-segment-group-context'
 
 export interface SegmentGroupLabelBaseProps extends PolymorphicProps {}
-export interface SegmentGroupLabelProps extends HTMLProps<'label'>, SegmentGroupLabelBaseProps {}
+export interface SegmentGroupLabelProps extends HTMLProps<'span'>, SegmentGroupLabelBaseProps {}
 
-export const SegmentGroupLabel = forwardRef<HTMLLabelElement, SegmentGroupLabelProps>((props, ref) => {
+export const SegmentGroupLabel = forwardRef<HTMLSpanElement, SegmentGroupLabelProps>((props, ref) => {
   const segmentGroup = useSegmentGroupContext()
   const mergedProps = mergeProps(segmentGroup.getLabelProps(), parts.label.attrs as Record<string, string>, props)
 
-  return <ark.label {...mergedProps} ref={ref} />
+  return <ark.span {...mergedProps} ref={ref} />
 })
 
 SegmentGroupLabel.displayName = 'SegmentGroupLabel'