+++ packages/solid/src/components/progress/progress-label.tsx
@@ -2,12 +2,12 @@ import { mergeProps } from '@zag-js/solid'
 import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { useProgressContext } from './use-progress-context'
 
-export interface ProgressLabelBaseProps extends PolymorphicProps<'label'> {}
-export interface ProgressLabelProps extends HTMLProps<'label'>, ProgressLabelBaseProps {}
+export interface ProgressLabelBaseProps extends PolymorphicProps<'span'> {}
+export interface ProgressLabelProps extends HTMLProps<'span'>, ProgressLabelBaseProps {}
 
 export const ProgressLabel = (props: ProgressLabelProps) => {
   const api = useProgressContext()
   const mergedProps = mergeProps(() => api().getLabelProps(), props)
 
-  return <ark.label {...mergedProps} />
+  return <ark.span {...mergedProps} />
 }