+++ packages/solid/src/components/segment-group/segment-group-label.tsx
@@ -3,12 +3,12 @@ import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { parts } from './segment-group.anatomy'
 import { useSegmentGroupContext } from './use-segment-group-context'
 
-export interface SegmentGroupLabelBaseProps extends PolymorphicProps<'label'> {}
-export interface SegmentGroupLabelProps extends HTMLProps<'label'>, SegmentGroupLabelBaseProps {}
+export interface SegmentGroupLabelBaseProps extends PolymorphicProps<'span'> {}
+export interface SegmentGroupLabelProps extends HTMLProps<'span'>, SegmentGroupLabelBaseProps {}
 
 export const SegmentGroupLabel = (props: SegmentGroupLabelProps) => {
   const segmentGroup = useSegmentGroupContext()
   const mergedProps = mergeProps(() => segmentGroup().getLabelProps(), parts.label.attrs, props)
 
-  return <ark.label {...mergedProps} />
+  return <ark.span {...mergedProps} />
 }