+++ packages/svelte/src/lib/components/segment-group/segment-group-label.svelte
@@ -1,8 +1,8 @@
 <script module lang="ts">
   import type { Assign, HTMLProps, PolymorphicProps, RefAttribute } from '$lib/types'
 
-  export interface SegmentGroupLabelBaseProps extends PolymorphicProps<'label'>, RefAttribute {}
-  export interface SegmentGroupLabelProps extends Assign<HTMLProps<'label'>, SegmentGroupLabelBaseProps> {}
+  export interface SegmentGroupLabelBaseProps extends PolymorphicProps<'span'>, RefAttribute {}
+  export interface SegmentGroupLabelProps extends Assign<HTMLProps<'span'>, SegmentGroupLabelBaseProps> {}
 </script>
 
 <script lang="ts">
@@ -17,4 +17,4 @@
   const mergedProps = $derived(mergeProps(segmentGroup().getLabelProps(), parts.label.attrs, props))
 </script>
 
-<Ark as="label" bind:ref {...mergedProps} />
+<Ark as="span" bind:ref {...mergedProps} />