+++ packages/vue/src/components/segment-group/segment-group-label.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import type { LabelHTMLAttributes } from 'vue'
+import type { HTMLAttributes } from 'vue'
 import type { PolymorphicProps } from '../factory'
 
 export interface SegmentGroupLabelBaseProps extends PolymorphicProps {}
@@ -9,7 +9,7 @@ export interface SegmentGroupLabelProps
     /**
      * @vue-ignore
      */
-    LabelHTMLAttributes {}
+    HTMLAttributes {}
 </script>
 
 <script setup lang="ts">
@@ -25,12 +25,12 @@ useForwardExpose()
 </script>
 
 <template>
-  <ark.label
+  <ark.span
     v-bind="segmentGroup.getLabelProps()"
     :as-child="asChild"
     :data-scope="parts.label.attrs['data-scope']"
     :data-part="parts.label.attrs['data-part']"
   >
     <slot />
-  </ark.label>
+  </ark.span>
 </template>